Combine main activity and activity fragment
TODO: fix rotation issue
